Tests/LibWeb/Text/expected/HTML/parser-streams-external-script.txt

script ran delayed script ran same chunk second chunk
